services/user-service/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
